requirements.txt
tensorflow==2.16.1
numpy
pillow
flask
gunicorn